Channels switching
Added a new table to the database called "messages". Table definition is (server, channel, user, timestamp, message). Each entry indicates a message and is intended to be used as a message log.

Channel switching is now complete in that a user can join a channel by clicking on the channel. If the channel is the one the user was in, the server does not move the user. The server will populate a list of messages for the new channel and send this message list to the user in order to populate their messages window. When a user switches channels, two broadcasts are done, one to the server, and one to the user switching. The user switching performs necessary UI updates to change their visual channel and messages, and users who receive the "user changing channel" broadcast update their UI to show the user in hte new channel. This is the only graphical change required client side for the other users.

I believe channel creation is already working (I forgot to mention it in prior comments I think), but there should be some cap on the number of channels a user can make. Need to decide if it's ideal that a user is allowed to delete channels or just rename them. Deletion shouldn't be too difficult because we will not allow the user to delete "DefaultChannel" and the database was designed with deletion in mind.