Added database script & got channels working. Channel text chat started.
Added a database script file that creates the SQL to create all tables in the schema. Added a startup line in the server that enables foreign key restrictions on the database. Retstrictions are checked manually
but foreign key support should be default.

User can add up to 10 unique channels to their server. Deletion has not yet been added, but will be added soon. Need to have a nicer dialog for asking for user input rather than javascript input windows.

Users can send chat right now, but it is limitted and not yet saved to database (for future connects to see channel history).