Persistent message history across channels
Messages are now stored in the database made in the previous commits.
Upon joining a channel, a user is sent the chat history of the channel which
is displayed in the client. Number of messages stored should be something
to think about. A history of more than 100 or so messages may not be necessary.

Also made it so that if users reconnect via refresh, that the server checks
to see if they were already in a channel in our stored history so that
duplicate names are not added to channels.